+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_creatersasigningkey.js
@@ -1,0 +1,5 @@
+let nock = require('nock');
+
+module.exports.hash = "3fc9b607de72d9c0c15ae925d17db4b5";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}