+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_createunsecuredattestationtoken.js
@@ -1,0 +1,5 @@
+let nock = require('nock');
+
+module.exports.hash = "9db64f0fbe9716f5174deab2b9c002d9";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}