+++ sdk/attestation/attestation/src/utils/textEncoding.browser.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+const textDecoder = TextDecoder;
+const textEncoder =  TextEncoder;
+
+export { textDecoder as TextDecoder, textEncoder as TextEncoder };