+++ sdk/datadog/arm-datadog/src/microsoftDatadogClientContext.ts
@@ -12,7 +12,7 @@ import * as msRest from "@azure/ms-rest-js";
 import * as msRestAzure from "@azure/ms-rest-azure-js";
 
 const packageName = "@azure/arm-datadog";
-const packageVersion = "1.0.0";
+const packageVersion = "2.0.0";
 
 export class MicrosoftDatadogClientContext extends msRestAzure.AzureServiceClient {
   credentials: msRest.ServiceClientCredentials;
@@ -25,12 +25,16 @@ export class MicrosoftDatadogClientContext extends msRestAzure.AzureServiceClien
    * @param subscriptionId The ID of the target subscription.
    * @param [options] The parameter options
    */
-  constructor(credentials: msRest.ServiceClientCredentials, subscriptionId: string, options?: Models.MicrosoftDatadogClientOptions) {
+  constructor(
+    credentials: msRest.ServiceClientCredentials,
+    subscriptionId: string,
+    options?: Models.MicrosoftDatadogClientOptions
+  ) {
     if (credentials == undefined) {
-      throw new Error('\'credentials\' cannot be null.');
+      throw new Error("'credentials' cannot be null.");
     }
     if (subscriptionId == undefined) {
-      throw new Error('\'subscriptionId\' cannot be null.');
+      throw new Error("'subscriptionId' cannot be null.");
     }
 
     if (!options) {
@@ -43,8 +47,8 @@ export class MicrosoftDatadogClientContext extends msRestAzure.AzureServiceClien
 
     super(credentials, options);
 
-    this.apiVersion = '2021-03-01';
-    this.acceptLanguage = 'en-US';
+    this.apiVersion = "2021-03-01";
+    this.acceptLanguage = "en-US";
     this.longRunningOperationRetryTimeout = 30;
     this.baseUri = options.baseUri || this.baseUri || "https://management.azure.com";
     this.requestContentType = "application/json; charset=utf-8";
@@ -54,7 +58,10 @@ export class MicrosoftDatadogClientContext extends msRestAzure.AzureServiceClien
     if (options.acceptLanguage !== null && options.acceptLanguage !== undefined) {
       this.acceptLanguage = options.acceptLanguage;
     }
-    if (options.longRunningOperationRetryTimeout !== null && options.longRunningOperationRetryTimeout !== undefined) {
+    if (
+      options.longRunningOperationRetryTimeout !== null &&
+      options.longRunningOperationRetryTimeout !== undefined
+    ) {
       this.longRunningOperationRetryTimeout = options.longRunningOperationRetryTimeout;
     }
   }