+++ sdk/core-rest/core-client/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release History
 
-## 1.0.0-beta.3 (Unreleased)
+## 1.0.0-beta.4 (2021-05-27)
+
+- Update @azure/core-rest-pipeline dependency to GA. [#15435](https://github.com/Azure/azure-sdk-for-js/pull/15435)
+
+## 1.0.0-beta.3 (2021-05-26)
+
+- Set Date in query string parameters as ISO string. [#15209](https://github.com/Azure/azure-sdk-for-js/pull/15209)
 
 ## 1.0.0-beta.2 (2021-05-12)