+++ sdk/videoanalyzer/video-analyzer-edge/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 1.0.0-beta.2 (2021-05-28)
+
+Update the title and description of the package and updates to readme. 
+
 ## 1.0.0-beta.1 (2021-05-20)
 
 Initial release